requirements.txt
accelerate==0.20.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
bitsandbytes==0.37.0
braceexpand==0.1.7
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decord==0.6.0
diffusers==0.23.0
distro==1.9.0
docker-pycreds==0.4.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.21.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
iopath==0.1.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
omegaconf==2.3.0
openai==1.59.8
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
peft==0.2.0
pillow==11.1.0
platformdirs==4.3.6
portalocker==3.1.1
progressbar2==4.5.0
protobuf==5.29.3
psutil==6.1.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==1.0.13
tokenizers==0.13.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.30.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
visual-genome==1.1.1
wandb==0.19.4
webdataset==0.2.100
zipp==3.21.0
